+++ social-network-backend/src/main/java/com/socialnetwork/project/service/UserService.java
@@ -2,5 +2,5 @@ package com.socialnetwork.project.service;
 
 import com.socialnetwork.project.entity.User;
 
-public interface UserService extends BaseService<User, Long> {
+public interface UserService extends BaseService<User, Integer> {
 }
